[Refactoring] Move LocalRename to its own file
